rename conda_fmt module to 'create'
