--- docs/conda_fmt.rst
@@ -1,7 +1,0 @@
-conda_fmt module
-================
-
-.. automodule:: conda_package_streaming.conda_fmt
-   :members:
-   :undoc-members:
-   :show-inheritance:
