+++ docs/create.md
@@ -1,0 +1,8 @@
+create module
+=============
+
+```.. automodule:: conda_package_streaming.create
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
