+++ docs/modules.md
@@ -78,5 +78,5 @@ lazy_wheel
 package_streaming
 extract
 transmute
-conda_fmt
+create
 ```
